refactor leetcode mock handlers into separate file
